# docs/swagger.yaml
basePath: /api
definitions:
...
  models.LoginResult:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserProfile'
    type: object
...
  models.UserProfile:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
      walletAddress:
        type: string
    type: object
  models.WalletLoginRequestNoncePayload:
    properties:
      walletAddress:
        type: string
    required:
    - walletAddress
    type: object
  models.WalletLoginRequestNonceResult:
    properties:
      message:
        type: string
      nonce:
        type: string
    type: object
  models.WalletLoginVerifyPayload:
    properties:
      message:
        type: string
      signature:
        type: string
      walletAddress:
        type: string
    required:
    - message
    - signature
    - walletAddress
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A RESTful API for NFT auction marketplace using Gin and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: My Auction Market API
  version: "1.0"
paths:
...
  /auth/wallet/request-nonce:
    post:
      consumes:
      - application/json
      description: Generates a unique nonce for a given wallet address to be signed
        by the user
      parameters:
      - description: Wallet address
        in: body
        name: walletAddress
        required: true
        schema:
          $ref: '#/definitions/models.WalletLoginRequestNoncePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.WalletLoginRequestNonceResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Request a nonce for wallet login
      tags:
      - auth
  /auth/wallet/verify:
    post:
      consumes:
      - application/json
      description: Verifies the signed message and logs in the user, returning a JWT
        token
      parameters:
      - description: Wallet login payload
        in: body
        name: walletLogin
        required: true
        schema:
          $ref: '#/definitions/models.WalletLoginVerifyPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Verify wallet signature and login
      tags:
      - auth
...
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get current user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"